博客分类页的展示博客分类标签名称与数量使用Mybatis重写

+++ src/main/java/com/dao/DetailDao.java
@@ -13,5 +13,5 @@ public interface DetailDao {
 
     List<BlogEntity> findAllBlogs();
 
-    List<BlogEntity> findAllBlogsByType(Long id);
+    List<BlogEntity> findAllBlogsByType(int id);
 }
